Move shared params
